+++ speechmatics/cli.py
@@ -35,7 +35,7 @@ from speechmatics.models import (
     BatchTranscriptionConfig,
     BatchSpeakerDiarizationConfig,
     RTSpeakerDiarizationConfig,
-    BatchTranslationConfig,
+    RTTranslationConfig,
     BatchLanguageIdentificationConfig,
 )
 from speechmatics.cli_parser import (
@@ -227,6 +227,8 @@ def get_transcription_config(args):  # pylint: disable=too-many-branches
     for option in [
         "enable_partials",
         "enable_entities",
+        "enable_translation_partials",
+        "enable_transcription_partials",
     ]:
         config[option] = True if args.get(option) else config.get(option)
 
@@ -274,8 +276,12 @@ def get_transcription_config(args):  # pylint: disable=too-many-branches
 
     if args.get("translation_target_languages") is not None:
         translation_target_languages = args.get("translation_target_languages")
-        config["translation_config"] = BatchTranslationConfig(
-            target_languages=translation_target_languages.split(",")
+        enable_partials = args.get("enable_partials", False) or args.get(
+            "enable_translation_partials", False
+        )
+        config["translation_config"] = RTTranslationConfig(
+            target_languages=translation_target_languages.split(","),
+            enable_partials=enable_partials,
         )
 
     if args.get("langid_expected_languages") is not None:
@@ -311,14 +317,17 @@ def get_audio_settings(args):
     return settings
 
 
-# pylint: disable=too-many-arguments
+# pylint: disable=too-many-arguments,too-many-statements
 def add_printing_handlers(
     api,
     transcripts,
     enable_partials=False,
+    enable_transcription_partials=False,
+    enable_translation_partials=False,
     debug_handlers_too=False,
     speaker_change_token=False,
     print_json=False,
+    translation_config=None,
 ):
     """
     Adds a set of handlers to the websocket client which print out transcripts
@@ -328,14 +337,23 @@ def add_printing_handlers(
         api (speechmatics.client.WebsocketClient): Client instance.
         transcripts (Transcripts): Allows the transcripts to be concatenated to
             produce a final result.
-        enable_partials (bool, optional): Whether or not partials are enabled.
-        debug_handlers_too (bool, optional): Whether or not to enable 'debug'
+        enable_partials (bool, optional): Whether partials are enabled
+            for both transcription and translation.
+        enable_transcription_partials (bool, optional): Whether partials are enabled
+            for transcription only.
+        enable_translation_partials (bool, optional): Whether partials are enabled
+            for translation only.
+        debug_handlers_too (bool, optional): Whether to enable 'debug'
             handlers that print out an ASCII symbol representing messages being
             received and sent.
         speaker_change_token (bool, optional): Whether to explicitly include a
             speaker change token '<sc>' in the output to indicate speaker
             changes.
+        print_json (bool, optional): Whether to print json transcript messages.
+        translation_config (TranslationConfig, optional): Translation config with target languages.
     """
+    escape_seq = "\33[2K" if sys.stdout.isatty() else ""
+
     if debug_handlers_too:
         api.add_event_handler(
             ServerMessageType.AudioAdded, lambda *args: print_symbol("-")
@@ -361,7 +379,6 @@ def add_printing_handlers(
             speaker_change_token=speaker_change_token,
         )
         if plaintext:
-            escape_seq = "\33[2K" if sys.stderr.isatty() else ""
             sys.stderr.write(f"{escape_seq}{plaintext}\r")
 
     def transcript_handler(message):
@@ -377,19 +394,69 @@ def add_printing_handlers(
             speaker_change_token=speaker_change_token,
         )
         if plaintext:
-            escape_seq = "\33[2K" if sys.stdout.isatty() else ""
             sys.stdout.write(f"{escape_seq}{plaintext}\n")
         transcripts.text += plaintext
+
+    def partial_translation_handler(message):
+        if print_json:
+            print(json.dumps(message))
+            return
+        # Translations for all requested languages should be available
+        # but, we're only going to print one translation
+        if translation_config.target_languages[0] == message["language"]:
+            plaintext = speechmatics.adapters.get_txt_translation(message["results"])
+            sys.stderr.write(f"{escape_seq}{plaintext}\r")
+
+    def translation_handler(message):
+        transcripts.json.append(message)
+        if print_json:
+            print(json.dumps(message))
+            return
+        # Translations for all requested languages should be available
+        # but, we're only going to print one translation
+        if translation_config.target_languages[0] == message["language"]:
+            plaintext = speechmatics.adapters.get_txt_translation(message["results"])
+            if plaintext:
+                sys.stdout.write(f"{escape_seq}{plaintext}\n")
+            transcripts.text += plaintext
 
     def end_of_transcript_handler(_):
         if enable_partials:
             print("\n", file=sys.stderr)
 
-    api.add_event_handler(
-        ServerMessageType.AddPartialTranscript, partial_transcript_handler
-    )
-    api.add_event_handler(ServerMessageType.AddTranscript, transcript_handler)
     api.add_event_handler(ServerMessageType.EndOfTranscript, end_of_transcript_handler)
+
+    # print both transcription and translation messages (if json was requested)
+    # print translation (if text was requested then)
+    # print transcription (if text was requested without translation)
+    if print_json:
+        if enable_partials or enable_translation_partials:
+            api.add_event_handler(
+                ServerMessageType.AddPartialTranslation,
+                partial_translation_handler,
+            )
+        api.add_event_handler(ServerMessageType.AddTranslation, translation_handler)
+        if enable_partials or enable_transcription_partials:
+            api.add_event_handler(
+                ServerMessageType.AddPartialTranscript,
+                partial_transcript_handler,
+            )
+        api.add_event_handler(ServerMessageType.AddTranscript, transcript_handler)
+    else:
+        if translation_config is not None:
+            if enable_partials or enable_translation_partials:
+                api.add_event_handler(
+                    ServerMessageType.AddPartialTranslation,
+                    partial_translation_handler,
+                )
+            api.add_event_handler(ServerMessageType.AddTranslation, translation_handler)
+        else:
+            if enable_partials or enable_transcription_partials:
+                api.add_event_handler(
+                    ServerMessageType.AddPartialTranscript,
+                    partial_transcript_handler,
+                )
+            api.add_event_handler(ServerMessageType.AddTranscript, transcript_handler)
 
 
 def join_words(words, language="en"):
@@ -456,7 +523,7 @@ def main(args=None):
     Main entrypoint.
 
     :param args: command-line arguments; defaults to None in which
-            case arguments will retrieved from `sys.argv` (this is useful
+            case arguments will be retrieved from `sys.argv` (this is useful
             mainly for unit tests).
     :type args: List[str]
     """
@@ -552,15 +619,18 @@ def rt_main(args):
         api,
         transcripts,
         enable_partials=args["enable_partials"],
+        enable_transcription_partials=args["enable_transcription_partials"],
+        enable_translation_partials=args["enable_translation_partials"],
         debug_handlers_too=args["debug"],
         speaker_change_token=args["speaker_change_token"],
         print_json=args["print_json"],
+        translation_config=transcription_config.translation_config,
     )
 
     def run(stream):
         try:
             api.run_synchronously(
-                stream, get_transcription_config(args), get_audio_settings(args)
+                stream, transcription_config, get_audio_settings(args)
             )
         except KeyboardInterrupt:
             # Gracefully handle Ctrl-C, else we get a huge stack-trace.