+++ speechmatics/client.py
@@ -49,6 +49,7 @@ class WebsocketClient:
         self.connection_settings = connection_settings
         self.websocket = None
         self.transcription_config = None
+        self.translation_config = None
 
         self.event_handlers = {x: [] for x in ServerMessageType}
         self.middlewares = {x: [] for x in ClientMessageType}
@@ -113,8 +114,10 @@ class WebsocketClient:
         """
         msg = {
             "message": ClientMessageType.SetRecognitionConfig,
-            "transcription_config": self.transcription_config.asdict(),
+            "transcription_config": self.transcription_config.as_config(),
         }
+        if self.translation_config is not None:
+            msg["translation_config"] = self.translation_config.asdict()
         self._call_middleware(ClientMessageType.SetRecognitionConfig, msg, False)
         return msg
 
@@ -132,8 +135,10 @@ class WebsocketClient:
         msg = {
             "message": ClientMessageType.StartRecognition,
             "audio_format": audio_settings.asdict(),
-            "transcription_config": self.transcription_config.asdict(),
+            "transcription_config": self.transcription_config.as_config(),
         }
+        if self.translation_config is not None:
+            msg["translation_config"] = self.translation_config.asdict()
         self.session_running = True
         self._call_middleware(ClientMessageType.StartRecognition, msg, False)
         LOGGER.debug(msg)
@@ -398,6 +403,7 @@ class WebsocketClient:
             consumer/producer tasks.
         """
         self.transcription_config = transcription_config
+        self.translation_config = transcription_config.translation_config
         self.seq_no = 0
         self._language_pack_info = None
         await self._init_synchronization_primitives()